+++ sql/addencryptedcols.sql
