+++ sql/create_DB.sql
@@ -1,0 +1,11 @@
+CREATE DATABASE "ClientDB"
+    WITH
+    OWNER = postgres
+    ENCODING = 'UTF8'
+    LC_COLLATE = 'English_United States.1252'
+    LC_CTYPE = 'English_United States.1252'
+    TABLESPACE = pg_default
+    CONNECTION LIMIT = -1
+    IS_TEMPLATE = False;
+
+ALTER DATABASE "ClientDB" OWNER TO postgres;
