+++ sql/dropcols.sql
